Create networkErrorHelpers.ts
