+++ src/errors/errorHelpers.ts
@@ -16,17 +16,15 @@
  */
 
 import { deserializeError, ErrorObject } from "serialize-error";
-import { AxiosError, AxiosResponse } from "axios";
 import { isObject, matchesAnyPattern } from "@/utils";
 import safeJsonStringify from "json-stringify-safe";
-import { isEmpty, truncate } from "lodash";
-import {
-  isBadRequestResponse,
-  isClientErrorResponse,
-  safeGuessStatusText,
-} from "@/types/errorContract";
+import { truncate } from "lodash";
 import type { ContextError } from "@/errors/genericErrors";
-import { Except } from "type-fest";
+import {
+  isAxiosError,
+  selectNetworkErrorMessage,
+  selectServerErrorMessage,
+} from "@/errors/networkErrorHelpers";
 
 const DEFAULT_ERROR_MESSAGE = "Unknown error";
 
@@ -152,24 +150,6 @@ export function isClientRequestError(error: unknown): boolean {
 }
 
 /**
- * Axios offers its own serialization method, but it doesn't include the response.
- * By deleting toJSON, the serialize-error library will use its default serialization
- */
-export type SerializableAxiosError = Except<AxiosError, "toJSON">;
-
-// Copy of axios.isAxiosError, without risking to import the whole untreeshakeable axios library
-export function isAxiosError(error: unknown): error is SerializableAxiosError {
-  if (isObject(error) && Boolean(error.isAxiosError)) {
-    // Axios offers its own serialization method, but it doesn't include the response.
-    // By deleting toJSON, the serialize-error library will use its default serialization
-    delete error.toJSON;
-    return true;
-  }
-
-  return false;
-}
-
-/**
  * Return true if the proximate cause of event is a messaging error.
  *
  * NOTE: does not recursively identify the root cause of the error.
@@ -194,105 +174,6 @@ export function isPrivatePageError(error: unknown): boolean {
   return /cannot be scripted|(chrome|about|extension):\/\//.test(
     getErrorMessage(error)
   );
-}
-
-export const NO_INTERNET_MESSAGE =
-  "No response received. You are not connected to the internet.";
-
-export const NO_RESPONSE_MESSAGE =
-  "No response received. Your browser may have blocked the request. See https://docs.pixiebrix.com/network-errors for troubleshooting information";
-
-function selectNetworkErrorMessage(error: unknown): string | null {
-  if (
-    (isAxiosError(error) && error.response == null) ||
-    (typeof (error as any).message === "string" &&
-      (error as { message: string }).message.toLowerCase() === "network error")
-  ) {
-    if (!navigator.onLine) {
-      return NO_INTERNET_MESSAGE;
-    }
-
-    return NO_RESPONSE_MESSAGE;
-  }
-
-  return null;
-}
-
-/**
- * Heuristically select the most user-friendly error message for an Axios response.
- *
- * Tries to handle:
- * - Errors produced by our backed (Django Rest Framework style)
- * - Common response body patterns of other APIs
- * - HTTP standards in statusText/status
- *
- * enrichBusinessRequestError is a related method which wraps an AxiosError in an Error subclass that encodes information
- * about why the request failed.
- *
- * @param response Response from the server. Must not be null
- *
- * @deprecated DO NOT CALL DIRECTLY. Call getErrorMessage
- * @see getErrorMessage
- * @see enrichBusinessRequestError
- */
-function selectServerErrorMessage(response: AxiosResponse): string | null {
-  if (response == null) {
-    throw new Error("Expected response to be defined");
-  }
-
-  // For examples of DRF errors, see the pixiebrix-app repository:
-  // http://github.com/pixiebrix/pixiebrix-app/blob/5ef1e4e414be6485fae999440b69f2b6da993668/api/tests/test_errors.py#L15-L15
-
-  // Handle 400 responses created by DRF serializers
-  if (isBadRequestResponse(response)) {
-    const data = Array.isArray(response.data)
-      ? response.data.find((x) => isEmpty(x))
-      : response.data;
-
-    // Prefer object-level errors
-    if (data?.non_field_errors) {
-      return data.non_field_errors[0];
-    }
-
-    // Take an arbitrary field
-    const fieldMessages = Object.values(data)[0];
-
-    // Take an arbitrary message
-    return fieldMessages[0];
-  }
-
-  // Handle 4XX responses created by DRF
-  if (isClientErrorResponse(response)) {
-    return response.data.detail;
-  }
-
-  // Handle other likely API JSON body response formats
-  // Some servers produce an HTML document for server responses even if you requested JSON. Check the response headers
-  // to avoid dumping JSON to the user
-  if (
-    typeof response.data === "string" &&
-    ["text/plain", "application/json"].includes(
-      response.headers["content-type"]
-    )
-  ) {
-    return response.data;
-  }
-
-  // Handle common keys from other APIs
-  for (const messageKey of ["message", "reason"]) {
-    // eslint-disable-next-line security/detect-object-injection -- constant defined above
-    const message = response.data?.[messageKey];
-    if (typeof message === "string" && !isEmpty(message)) {
-      return message;
-    }
-  }
-
-  // Otherwise, rely on HTTP REST statusText/status
-  if (!isEmpty(response.statusText)) {
-    return response.statusText;
-  }
-
-  return safeGuessStatusText(response.status);
 }
 
 /**
