+++ src/hooks/fetch.ts
@@ -20,8 +20,8 @@ import { getApiClient, getLinkedApiClient } from "@/services/apiClient";
 import { isAbsoluteUrl } from "@/utils";
 import { isAppUrl } from "@/services/requestErrorUtils";
 import { expectContext, forbidContext } from "@/utils/expectContext";
-import { isAxiosError } from "@/errors/errorHelpers";
 import { EndpointAuthError } from "@/errors/genericErrors";
+import { isAxiosError } from "@/errors/networkErrorHelpers";
 
 const HTTP_401_UNAUTHENTICATED = 401;
