+++ src/options/pages/blueprints/modals/ConvertToRecipeModal.tsx
@@ -47,7 +47,7 @@ import { FieldDescriptions } from "@/utils/strings";
 import RegistryIdWidget from "@/components/form/widgets/RegistryIdWidget";
 import { StylesConfig } from "react-select";
 import { RequireScope } from "@/auth/RequireScope";
-import { isSingleObjectBadRequestError } from "@/types/errorContract";
+import { isSingleObjectBadRequestError } from "@/errors/networkErrorHelpers";
 
 type ConvertInstallableFormState = {
   blueprintId: RegistryId;
