+++ src/pageEditor/sidebar/CreateRecipeModal.tsx
@@ -66,7 +66,7 @@ import useRemoveExtension from "@/pageEditor/hooks/useRemoveExtension";
 import useRemoveRecipe from "@/pageEditor/hooks/useRemoveRecipe";
 import RegistryIdWidget from "@/components/form/widgets/RegistryIdWidget";
 import { generateRecipeId } from "@/utils/recipeUtils";
-import { isSingleObjectBadRequestError } from "@/types/errorContract";
+import { isSingleObjectBadRequestError } from "@/errors/networkErrorHelpers";
 
 const { actions: optionsActions } = extensionsSlice;
