+++ src/services/requestErrorUtils.ts
@@ -15,11 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import {
-  isAxiosError,
-  isErrorObject,
-  SerializableAxiosError,
-} from "@/errors/errorHelpers";
+import { isErrorObject } from "@/errors/errorHelpers";
 import { AxiosRequestConfig } from "axios";
 import { testMatchPatterns } from "@/blocks/available";
 import {
@@ -29,6 +25,10 @@ import {
 } from "@/services/baseService";
 import { isAbsoluteUrl } from "@/utils";
 import urljoin from "url-join";
+import {
+  isAxiosError,
+  SerializableAxiosError,
+} from "@/errors/networkErrorHelpers";
 
 /**
  * Get the absolute URL from a request configuration. Does NOT include the query params from the request unless
